RTE: Rename div-container extension file for div- prefix grouping
Renames html-tag-container-div.tiptap-extension.ts to
html-tag-div-container.tiptap-extension.ts so it groups with the
html-tag-div-* prefix instead of a container-div suffix.
