+++ src/Umbraco.Web.UI.Client/src/packages/tiptap/extensions/html-tag-div/html-tag-div.tiptap-api.ts
@@ -1,6 +1,6 @@
 import { UmbTiptapExtensionApiBase } from '../tiptap-extension-api-base.js';
 import { Div } from './html-tag-div.tiptap-extension.js';
-import { DivContainer } from './html-tag-container-div.tiptap-extension.js';
+import { DivContainer } from './html-tag-div-container.tiptap-extension.js';
 
 export default class UmbTiptapHtmlTagDivExtensionApi extends UmbTiptapExtensionApiBase {
 	getTiptapExtensions = () => [Div, DivContainer];
